Added Screen Switching (Part 2)
Added the ability to go forward and back with the screen. Due to how the mouse clicker works, clicking on the UI elements still triggers the script. This is an issue, as if there is no object behind the UI element the mouse will just treat it as you clicking on nothing. This isn't a big deal, but it might be something worth fixing for later.