Added Spacebar Button and UI Blocking for Mouse
Added the ability for the UI to block the player's mouse for mouse interactions with the scene. I also added the space bar for going forward into the scene. I'm going to need to setup a way for the space bar behaviour to change when the player is locked in a puzzle.