Disabling Inactive Rooms
Made it so that the rooms are only active when the player is in them. This causes a problem though since I can't do it in the start function, so I have to do it at the start of the first update. The screens also have the option to disable the room upon leaving a room. Need to streamline this.