Made Some Small Fixes - Note of Exploit
Just made some small fixes, but I also have a note to mention regarding the screen lock.

There is an exploit with the screen lock. If you click on a locked screen the screen will be inaccessible, but the object will still be saved. This is because it was still the lastClicked object. If you then unlock the screen and shift screens (but not to the newly unlocked screen) it will still be available as the forward screen. This is because it is still the last clicked object, and only goes away once the user clicks on something else. This exploit likely won't come up in practice, but it's still something that should be fixed.