Final Accessibility Push For Capstone
The accessibility is almost fully implemented. Due to some errors, lightbox solutions, activating objects on the closet, the conveyor belt buttons, and the maze do not work using accessibility, but all can be activated with the mouse while using the screen reader.

The screen reader implementation is there, it just needs to be changed so it works properly, but I haven't been able to figure out how.